Added the API for cancel loan and create loan API changes
git-svn-id: http://192.168.110.32/svn/Products/PFF/core/branches/mlod@110706 6963f9ea-81fa-44da-aa10-7ebde50aa541

+++ PFSJava/src/main/java/com/pennant/backend/service/finance/impl/FinanceCancellationServiceImpl.java
@@ -534,7 +534,7 @@ public class FinanceCancellationServiceImpl extends GenericFinanceDetailService 
 		}
 
 		// Adding audit as deleted from TEMP table
-		if (isNotReqEOD) {
+		if (isNotReqEOD && auditHeader.getApiHeader() == null) {
 			getFinanceMainDAO().delete(financeMain, TableType.TEMP_TAB, false, true);
 		}
 		auditHeader.setAuditDetail(new AuditDetail(auditHeader.getAuditTranType(), 1, fields[0], fields[1],
